refactor(commands): modularize commands into separate files for better maintainability
The `commands.ts` file was split into multiple smaller files (`autostart.ts`, `config.ts`, `monitor.ts`, `system.ts`, `theme.ts`, `translation.ts`, `window.ts`) to improve code organization and maintainability. This change makes it easier to manage and extend individual command functionalities in the future.
