+++ src/commands/index.ts
@@ -1,0 +1,7 @@
+export * from "./autostart";
+export * from "./config";
+export * from "./monitor";
+export * from "./system";
+export * from "./theme";
+export * from "./translation";
+export * from "./window";
