+++ src/commands/monitor.ts
@@ -1,0 +1,4 @@
+import { invoke } from "@tauri-apps/api/core";
+
+export const getMonitors = async () =>
+  invoke<Record<string, MonitorInfo>>("get_monitors");
