+++ src/commands/translation.ts
@@ -1,0 +1,4 @@
+import { invoke } from "@tauri-apps/api/core";
+
+export const getTranslations = async () =>
+  invoke<Record<TranslationKey, string>>("get_translations");
